+++ tests/image_test.py
@@ -280,6 +280,19 @@ def test_set_units(fx_asset):
         assert img.units == "pixelspercentimeter"
 
 
+def test_get_colorspace(fx_asset):
+    """Gets the image colorspace"""
+    with Image(filename=str(fx_asset.join('mona-lisa.jpg'))) as img:
+        assert img.colorspace.endswith('rgb')
+
+
+def test_set_colorspace(fx_asset):
+    """Sets the image colorspace"""
+    with Image(filename=str(fx_asset.join('mona-lisa.jpg'))) as img:
+        img.colorspace = 'cmyk'
+        assert img.colorspace == 'cmyk'
+
+
 def test_get_depth(fx_asset):
     """Gets the image depth"""
     with Image(filename=str(fx_asset.join('mona-lisa.jpg'))) as img:
@@ -611,25 +624,33 @@ def test_crop_error(fx_asset):
             img.crop(bottom=1, height=2)
 
 
-def test_resize(fx_asset):
-    """Resizes the image."""
+@mark.parametrize(('method'), [
+    ('resize'),
+    ('sample'),
+])
+def test_resize_and_sample(method, fx_asset):
+    """Resizes/Samples the image."""
     with Image(filename=str(fx_asset.join('mona-lisa.jpg'))) as img:
         with img.clone() as a:
             assert a.size == (402, 599)
-            a.resize(100, 100)
+            getattr(a, method)(100, 100)
             assert a.size == (100, 100)
         with img.clone() as b:
             assert b.size == (402, 599)
-            b.resize(height=100)
+            getattr(b, method)(height=100)
             assert b.size == (402, 100)
         with img.clone() as c:
             assert c.size == (402, 599)
-            c.resize(width=100)
+            getattr(c, method)(width=100)
             assert c.size == (100, 599)
 
 
 @mark.slow
-def test_resize_gif(tmpdir, fx_asset):
+@mark.parametrize(('method'), [
+    ('resize'),
+    ('sample'),
+])
+def test_resize_and_sample_gif(method, tmpdir, fx_asset):
     with Image(filename=str(fx_asset.join('nocomments-delay-100.gif'))) as img:
         assert len(img.sequence) == 46
         with img.clone() as a:
@@ -637,7 +658,7 @@ def test_resize_gif(tmpdir, fx_asset):
             assert a.sequence[0].delay == 100
             for s in a.sequence:
                 assert s.delay == 100
-            a.resize(175, 98)
+            getattr(a, method)(175, 98)
             a.save(filename=str(tmpdir.join('175_98.gif')))
         with Image(filename=str(tmpdir.join('175_98.gif'))) as a:
             assert len(a.sequence) == 46
@@ -648,7 +669,7 @@ def test_resize_gif(tmpdir, fx_asset):
             assert b.size == (350, 197)
             for s in b.sequence:
                 assert s.delay == 100
-            b.resize(height=100)
+            getattr(b, method)(height=100)
             b.save(filename=str(tmpdir.join('350_100.gif')))
         with Image(filename=str(tmpdir.join('350_100.gif'))) as b:
             assert len(b.sequence) == 46
@@ -659,7 +680,7 @@ def test_resize_gif(tmpdir, fx_asset):
             assert c.size == (350, 197)
             for s in c.sequence:
                 assert s.delay == 100
-            c.resize(width=100)
+            getattr(c, method)(width=100)
             c.save(filename=str(tmpdir.join('100_197.gif')))
         with Image(filename=str(tmpdir.join('100_197.gif'))) as c:
             assert len(c.sequence) == 46
@@ -669,21 +690,25 @@ def test_resize_gif(tmpdir, fx_asset):
     tmpdir.remove()
 
 
-def test_resize_errors(fx_asset):
-    """Resizing errors."""
+@mark.parametrize(('method'), [
+    ('resize'),
+    ('sample'),
+])
+def test_resize_and_sample_errors(method, fx_asset):
+    """Resizing/Sampling errors."""
     with Image(filename=str(fx_asset.join('mona-lisa.jpg'))) as img:
         with raises(TypeError):
-            img.resize(width='100')
+            getattr(img, method)(width='100')
         with raises(TypeError):
-            img.resize(height='100')
+            getattr(img, method)(height='100')
         with raises(ValueError):
-            img.resize(width=0)
+            getattr(img, method)(width=0)
         with raises(ValueError):
-            img.resize(height=0)
+            getattr(img, method)(height=0)
         with raises(ValueError):
-            img.resize(width=-5)
+            getattr(img, method)(width=-5)
         with raises(ValueError):
-            img.resize(height=-5)
+            getattr(img, method)(height=-5)
 
 
 @mark.parametrize(('args', 'kwargs', 'expected_size'), [
@@ -1197,3 +1222,25 @@ def test_gaussian_blur(fx_asset, display):
         assert 0.84 <= after.red <= 0.85
         assert 0.74 <= after.green <= 0.75
         assert 0.655 <= after.blue < 0.67
+
+
+def test_modulate(fx_asset, display):
+    with Image(filename=str(fx_asset.join('sasha.jpg'))) as img:
+        before = img[100,100]
+        img.modulate(120, 120, 120)
+        after = img[100,100]
+        assert before != after
+        assert 0.98 <= after.red <= 0.99
+        assert 0.98 <= after.green <= 0.99
+        assert 0.96 <= after.blue <= 0.97
+
+
+def test_unsharp_mask(fx_asset, display):
+    with Image(filename=str(fx_asset.join('sasha.jpg'))) as img:
+        before = img[100, 100]
+        img.unsharp_mask(1.1, 1, 0.5, 0.001)
+        after = img[100, 100]
+        assert before != after
+        assert 0.89 <= after.red <= 0.90
+        assert 0.82 <= after.green <= 0.83
+        assert 0.73 <= after.blue < 0.74